move module editor to separate folder
